requirements.txt
requests
pandas
python-dotenv
supabase
psycopg2-binary
streamlit
plotly
